Homepage frontend methods

+++ Backend/Games/Keno/KenoOdds.cs
@@ -3,7 +3,7 @@
     public class KenoOdds
     {
         public int PlayerNumbers { get; set; }
-        public double Multiplier { get; set; }
+        public decimal Multiplier { get; set; }
 
     }
 }
